角色管理功能配置权限菜单功能修改完善

+++ src/main/java/com/tnpy/mes/mapper/mysql/TbRoleMenuMapper.java
@@ -17,7 +17,7 @@ public interface TbRoleMenuMapper
      * @param roleId 角色ID
      * @return 结果
      */
-    public int deleteRoleMenuByRoleId(Long roleId);
+    public int deleteRoleMenuByRoleId(Integer roleId);
     
     /**
      * 批量删除角色菜单关联信息
